Initial commit.
Created function that turns of pixels in an image that are not the local maximum in the intersection of the edge.

Created a function that turns off pixels below a certain threshold a, and keeps pixels on above threshold b. Pixels between the two thresholds are only kept on if they are connected (directly or indirectly) with a pixel that passes threshold b.